// src/types/dashboard.ts
export interface IDashboard {
    totalPatients: number
    fullyVaccinated: number
    partiallyVaccinated: number
    notVaccinated: number
    mumbaiPatients: number
    banglorePatients: number
    covishieldPatients: number
    covaxinPatients: number
    malePatients: number
    femalePatients: number
    otherPatients: number
}